moving lib and custom ui to lib and ui
