+++ app/dashboard/page.tsx
@@ -1,10 +1,10 @@
 'use client';
-import CustomerDetails from "../ui/Dashboard/Customer/MainPage";
-import OwnerDetails from "../ui/Dashboard/Owner/MainPage";
-import AdminDetails from "../ui/Dashboard/Admin/MainPage";
-import { useAuthStore } from "../lib/store";
+import CustomerDetails from "@/components/custom-ui/Dashboard/Customer/MainPage";
+import OwnerDetails from "@/components/custom-ui/Dashboard/Owner/MainPage";
+import AdminDetails from "@/components/custom-ui/Dashboard/Admin/MainPage";
+import { useAuthStore } from "../../lib/store";
 import { ProtectedRoute } from "../protector";
-import { allRoles } from "../lib/utils";
+import {allRoles} from "@/lib/utils";
 
 
 export default function Dashboard() {
