+++ components/custom-ui/Dashboard/Admin/MainPage.tsx
@@ -1,14 +1,15 @@
 'use client';
-import { getLoggedUserProfile } from "@/app/lib/service";
+import { getLoggedUserProfile } from "@/lib/service";
 import { toast } from "@/hooks/use-toast";
 import { useEffect, useState } from "react";
-import { UserDetails } from "@/app/lib/definitions";
+import { UserDetails } from "@/lib/definitions";
 import { Skeleton } from "@/components/ui/skeleton";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { AlertCircle, User } from "lucide-react";
-import { Orders } from "./orderList";
+import { UserList } from "./UserList";
+import AllBookings from "./bookings";
 
-export default function OwnerDetails() {
+export default function AdminDetails() {
   const [userDetails, setUserDetails] = useState<UserDetails>();
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState<string | null>();
@@ -65,11 +66,12 @@ export default function OwnerDetails() {
             Welcome back, {userDetails.fullName}
           </h1>
           <p className="text-muted-foreground">
-            Manage your machine bookings and view your history
+            Manage your users and everything
           </p>
         </div>
       </div>
-      <Orders userDetails={userDetails} ownerId={String(userDetails.id)} />
+      <AllBookings />
+      <UserList />
     </div>
   );
 }
