+++ components/custom-ui/Dashboard/Customer/bookingstable.tsx
@@ -1,5 +1,5 @@
-import { BookingListResponse } from "@/app/lib/definitions";
-import { getBookingsByUser, getLoggedUserProfile } from "@/app/lib/service";
+import { BookingListResponse } from "@/lib/definitions";
+import { getBookingsByUser, getLoggedUserProfile } from "@/lib/service";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Badge } from "@/components/ui/badge";
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
