+++ components/custom-ui/HomePage/featuredEquipment.tsx
