+++ components/custom-ui/login.tsx
@@ -12,7 +12,7 @@ import {
     FormMessage,
 } from "@/components/ui/form"
 import { Input } from "@/components/ui/input"
-import { LoginRequest } from "../lib/definitions"
+import { LoginRequest } from "../../lib/definitions"
 import Link from "next/link"
 
 
@@ -99,7 +99,7 @@ export function LoginForm(
                 </div>
                 
                 <div className="p-2">
-                    <Link href="/" >Go back home </Link>
+                    <Link href="/public" >Go back home </Link>
                 </div>
             </div>
