+++ components/custom-ui/machines/machineForm.tsx
@@ -11,8 +11,8 @@ import {
 import { Input } from "@/components/ui/input";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
 import { Textarea } from "@/components/ui/textarea";
-import { CategoryListResponse, MachineRequest } from "@/app/lib/definitions";
-import { getAllCategories, getLoggedUserProfile, getMachineConditions } from "@/app/lib/service";
+import { CategoryListResponse, MachineRequest } from "@/lib/definitions";
+import { getAllCategories, getLoggedUserProfile, getMachineConditions } from "@/lib/service";
 import { toast } from "@/hooks/use-toast";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useEffect, useState } from "react";
